--- pages/show.vue
@@ -1,9 +1,0 @@
-<template>
-  <Tutorial />
-</template>
-
-<script>
-export default {
-  name: 'ShowPage',
-}
-</script>
